SampleDesign totally breaking now in this commit.  Saving to come back to
