+++ src/webparts/v2NpmBanner/components/HelpPanel/About.tsx
@@ -6,7 +6,7 @@ import { repoLink, } from '../../fpsReferences';
 
 import { createAboutRow } from '../../fpsReferences';
 
-export const panelVersionNumber = '2022-10-24 -  1.0.0.05'; //Added to show in panel
+export const panelVersionNumber = '2022-10-23 -  1.0.0.05'; //Added to show in panel
 
 export function aboutTable( showRepoLinks: boolean ): { table: IHelpTable; } {
 
@@ -16,15 +16,15 @@ export function aboutTable( showRepoLinks: boolean ): { table: IHelpTable; } {
         rows: [],
     };
 
-    table.rows.push( createAboutRow('2022-10-14',"1.0.0.05","#30, #31, #32, #33, #34, #35", showRepoLinks === true ? repoLink : null ) );
-    table.rows.push( createAboutRow('2022-10-14',"1.0.0.05","#23, #24, #26, #27, #28, #29,", showRepoLinks === true ? repoLink : null ) );
+    table.rows.push( createAboutRow('2022-10-23',"1.0.0.05","#30, #31, #32, #33, #34, #35", showRepoLinks === true ? repoLink : null ) );
+    table.rows.push( createAboutRow('',"","#23, #24, #25, #26, #27, #28, #29,", showRepoLinks === true ? repoLink : null ) );
 
-    table.rows.push( createAboutRow('2022-10-14',"1.0.0.04","#12, #13, #14, #15, #17, #18, #19, #20, #22", showRepoLinks === true ? repoLink : null ) );
-    table.rows.push( createAboutRow('2022-10-14',"1.0.0.03","#3, #4, #6, #8, #9, #10", showRepoLinks === true ? repoLink : null ) );
+    table.rows.push( createAboutRow('2022-10-19',"1.0.0.04","#12, #13, #14, #15, #17, #18, #19, #20, #22", showRepoLinks === true ? repoLink : null ) );
+    table.rows.push( createAboutRow('2022-10-18',"1.0.0.03","#3, #4, #6, #8, #9, #10", showRepoLinks === true ? repoLink : null ) );
     table.rows.push( createAboutRow('2022-10-14',"1.0.0.02","#2,", showRepoLinks === true ? repoLink : null ) );
     table.rows.push( createAboutRow('2022-10-13',"1.0.0.01","Update _performance in main react component", showRepoLinks === true ? repoLink : null ) );
     table.rows.push( createAboutRow('2022-10-13',"1.0.0.00","Initial Build", showRepoLinks === true ? repoLink : null ) );
-    
+
     return { table: table };
 
 }
